Fix some compiler warnings.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@2434 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ sc_url_parse.cpp
@@ -203,7 +203,7 @@ void SaveCache()
   }
   //mark expired and count all
   unsigned int n=0;
-  double nowTime= time( NULL );
+  double nowTime= (double) time( NULL );
   for ( size_t i=1; i<=N_cache; i++ )
   {
     bool expired = nowTime - urlCache[i].time > expirationSeconds;
@@ -297,7 +297,7 @@ std::string getURLData( std::string URL )
         found=i;
         break;
       }
-  double nowTime= time( NULL );
+  double nowTime= (double) time( NULL );
   bool expired = ( !found ) || ( nowTime-urlCache[found].time > expirationSeconds );
   if ( expired )
   {
